+++ common/src/main/java/com/ss/android/ugc/bytex/common/CommonTransform.java
@@ -5,7 +5,8 @@ import com.android.build.api.transform.SecondaryFile;
 import com.android.build.api.transform.Transform;
 import com.android.build.api.transform.TransformException;
 import com.android.build.api.transform.TransformInvocation;
-import com.android.build.gradle.internal.pipeline.TransformManager;
+import com.android.build.api.variant.VariantInfo;
+import com.android.build.gradle.internal.scope.VariantScope;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
@@ -17,8 +18,10 @@ import com.ss.android.ugc.bytex.common.internal.TransformFlowerManager;
 import com.ss.android.ugc.bytex.common.log.LevelLog;
 import com.ss.android.ugc.bytex.common.log.Timer;
 import com.ss.android.ugc.bytex.common.log.html.HtmlReporter;
-import com.ss.android.ugc.bytex.gradletoolkit.TransformInvocationKt;
+import com.ss.android.ugc.bytex.common.utils.HeapDumper;
+import com.ss.android.ugc.bytex.gradletoolkit.ProjectKt;
 import com.ss.android.ugc.bytex.transformer.TransformContext;
+import com.ss.android.ugc.bytex.transformer.TransformOptions;
 
 import java.io.File;
 import java.io.IOException;
@@ -30,6 +33,8 @@ import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import javax.annotation.Nullable;
+
 /**
  * Created by tlh on 2018/8/29.
  */
@@ -37,6 +42,9 @@ import java.util.stream.Collectors;
 public abstract class CommonTransform<X extends BaseContext> extends Transform {
     protected final X context;
     private Set<TransformConfiguration> configurations;
+    @Nullable
+    //3.4在配置阶段才有
+    private String applyingVariantName = null;
 
     public CommonTransform(X context) {
         this.context = context;
@@ -45,6 +53,17 @@ public abstract class CommonTransform<X extends BaseContext> extends Transform {
     @Override
     public String getName() {
         return context.extension.getName();
+    }
+
+    @Override
+    public final boolean applyToVariant(VariantInfo variant) {
+        applyingVariantName = variant.getFullVariantName();
+        return super.applyToVariant(variant);
+    }
+
+    @Nullable
+    private VariantScope getApplyingVariantScope() {
+        return applyingVariantName == null ? null : ProjectKt.findVariantScope(context.project, applyingVariantName);
     }
 
     @Override
@@ -57,7 +76,7 @@ public abstract class CommonTransform<X extends BaseContext> extends Transform {
             }
         }
         if (result.isEmpty()) {
-            return TransformManager.CONTENT_CLASS;
+            return TransformConfiguration.DEFAULT.getInputTypes();
         }
         return result;
     }
@@ -65,14 +84,15 @@ public abstract class CommonTransform<X extends BaseContext> extends Transform {
     @Override
     public Set<? super QualifiedContent.Scope> getScopes() {
         Set<? super QualifiedContent.Scope> result = ImmutableSet.of();
+        VariantScope variantScope = getApplyingVariantScope();
         for (TransformConfiguration config : getConfigurations()) {
-            Set<? super QualifiedContent.Scope> scopes = config.getScopes();
+            Set<? super QualifiedContent.Scope> scopes = config.getScopes(variantScope);
             if (!result.containsAll(scopes)) {
                 result = Sets.union(result, scopes);
             }
         }
         if (result.isEmpty()) {
-            return TransformManager.SCOPE_FULL_PROJECT;
+            return TransformConfiguration.DEFAULT.getScopes(variantScope);
         }
         return result;
     }
@@ -87,7 +107,7 @@ public abstract class CommonTransform<X extends BaseContext> extends Transform {
             }
         }
         if (result.isEmpty()) {
-            return TransformManager.CONTENT_CLASS;
+            return TransformConfiguration.DEFAULT.getOutputTypes();
         }
         return result;
     }
@@ -95,14 +115,15 @@ public abstract class CommonTransform<X extends BaseContext> extends Transform {
     @Override
     public Set<? super QualifiedContent.Scope> getReferencedScopes() {
         Set<? super QualifiedContent.Scope> result = super.getReferencedScopes();
+        VariantScope variantScope = getApplyingVariantScope();
         for (TransformConfiguration config : getConfigurations()) {
-            Set<? super QualifiedContent.Scope> referencedScopes = config.getReferencedScopes();
+            Set<? super QualifiedContent.Scope> referencedScopes = config.getReferencedScopes(variantScope);
             if (!result.containsAll(referencedScopes)) {
                 result = Sets.union(result, referencedScopes);
             }
         }
         if (result.isEmpty()) {
-            return TransformManager.SCOPE_FULL_PROJECT;
+            return TransformConfiguration.DEFAULT.getReferencedScopes(variantScope);
         }
         return result;
     }
@@ -207,6 +228,9 @@ public abstract class CommonTransform<X extends BaseContext> extends Transform {
             }
             afterTransform(transformInvocation);
         } catch (Throwable throwable) {
+            if (throwable instanceof OutOfMemoryError) {
+                HeapDumper.dumpHeap(context.project.getBuildDir().getAbsolutePath() + "/oom.hprof", true);
+            }
             LevelLog.sDefaultLogger.e(throwable.getClass().getName(), throwable);
             throw throwable;
         } finally {
@@ -229,7 +253,17 @@ public abstract class CommonTransform<X extends BaseContext> extends Transform {
     }
 
     protected TransformContext getTransformContext(TransformInvocation transformInvocation) {
-        return new TransformContext(transformInvocation, context.project, context.android, isIncremental(), shouldSaveCache(), BooleanProperty.ENABLE_RAM_CACHE.value());
+        return new TransformContext(transformInvocation,
+                context.project,
+                context.android,
+                new TransformOptions.Builder()
+                        .setPluginIncremental(isIncremental())
+                        .setShouldSaveCache(shouldSaveCache())
+                        .setUseRawCache(BooleanProperty.ENABLE_RAM_CACHE.value())
+                        .setUseFixedTimestamp(BooleanProperty.USE_FIXED_TIMESTAMP.value())
+                        .setForbidUseLenientMutationDuringGetArtifact(BooleanProperty.FORBID_USE_LENIENT_MUTATION_DURING_GET_ARTIFACT.value())
+                        .build()
+        );
     }
 
     protected void afterTransform(TransformInvocation transformInvocation) throws TransformException, InterruptedException, IOException {
@@ -241,7 +275,7 @@ public abstract class CommonTransform<X extends BaseContext> extends Transform {
             String applicationId = "unknow";
             String versionName = "unknow";
             String versionCode = "unknow";
-            com.android.builder.model.ProductFlavor flavor = TransformInvocationKt.getVariant(transformContext.getInvocation()).getMergedFlavor();
+            com.android.builder.model.ProductFlavor flavor = transformContext.getVariant().getMergedFlavor();
             if (flavor != null) {
                 String flavorApplicationId = flavor.getApplicationId();
                 if (flavorApplicationId != null && !flavorApplicationId.isEmpty()) {