+++ refer-check-plugin/src/main/java/com/ss/android/ugc/bytex/refercheck/visitor/ReferCheckMethodVisitor.java
@@ -1,5 +1,6 @@
 package com.ss.android.ugc.bytex.refercheck.visitor;
 
+import com.ss.android.ugc.bytex.common.Constants;
 import com.ss.android.ugc.bytex.common.graph.ClassNode;
 import com.ss.android.ugc.bytex.common.graph.FieldEntity;
 import com.ss.android.ugc.bytex.common.graph.Graph;
@@ -7,10 +8,10 @@ import com.ss.android.ugc.bytex.common.graph.InterfaceNode;
 import com.ss.android.ugc.bytex.common.graph.MemberEntity;
 import com.ss.android.ugc.bytex.common.graph.MethodEntity;
 import com.ss.android.ugc.bytex.common.graph.Node;
+import com.ss.android.ugc.bytex.common.utils.MethodMatcher;
 import com.ss.android.ugc.bytex.common.utils.TypeUtil;
 import com.ss.android.ugc.bytex.common.utils.Utils;
 import com.ss.android.ugc.bytex.refercheck.InaccessibleNode;
-import com.ss.android.ugc.bytex.refercheck.ReferCheckContext;
 
 import org.objectweb.asm.Label;
 import org.objectweb.asm.MethodVisitor;
@@ -18,30 +19,39 @@ import org.objectweb.asm.Opcodes;
 
 import java.util.List;
 
+import javax.annotation.Nullable;
+
 /**
  * Created by tlh on 2018/2/13.
  */
 
 public class ReferCheckMethodVisitor extends MethodVisitor {
-    private final ReferCheckContext context;
+    private final CheckIssueReceiver checkIssueReceiver;
     private final String sourceFile;
     private final Graph graph;
+    private final List<MethodMatcher> blockMethodListMatchers;
     private final String className;
     private final String methodName;
     private final String methodDesc;
     private final int methodAccess;
     private int processingLineNumber;
 
-    ReferCheckMethodVisitor(MethodVisitor mv, ReferCheckContext context,
+    ReferCheckMethodVisitor(MethodVisitor mv, CheckIssueReceiver checkIssueReceiver, Graph graph,
+                            List<MethodMatcher> blockMethodListMatchers,
                             String className, String methodName, String methodDesc, int methodAccess, String sourceFile) {
-        super(Opcodes.ASM5, mv);
-        this.context = context;
+        super(Constants.ASM_API, mv);
+        this.checkIssueReceiver = checkIssueReceiver;
         this.className = className;
         this.methodName = methodName;
         this.methodDesc = methodDesc;
         this.methodAccess = methodAccess;
         this.sourceFile = sourceFile;
-        this.graph = context.getClassGraph();
+        this.graph = graph;
+        this.blockMethodListMatchers = blockMethodListMatchers;
+        if (TypeUtil.isAbstract(methodAccess)) {
+            boolean itf = graph.get(className) instanceof InterfaceNode;
+            checkMethod(itf ? Opcodes.INVOKEINTERFACE : Opcodes.INVOKEVIRTUAL, className, methodName, methodDesc, itf);
+        }
     }
 
     @Override
@@ -52,16 +62,19 @@ public class ReferCheckMethodVisitor extends MethodVisitor {
 
     @Override
     public void visitMethodInsn(int opcode, String owner, String name, String desc, boolean itf) {
+        for (MethodMatcher blockMethodListMatcher : blockMethodListMatchers) {
+            if (blockMethodListMatcher.match(owner, name, desc)) {
+                checkIssueReceiver.addNotAccessMember(
+                        className, methodName, methodDesc, methodAccess, sourceFile, processingLineNumber,
+                        owner, name, desc, 0, InaccessibleNode.TYPE_CALL_BLOCK_METHOD);
+                break;
+            }
+        }
         checkMethod(opcode, owner, name, desc, itf);
         super.visitMethodInsn(opcode, owner, name, desc, itf);
     }
 
     private void checkMethod(final int opcode, final String owner, final String name, final String desc, final boolean itf) {
-        String method = String.format("%s#%s", owner, name);
-        List<String> logMethods = context.extension.getLogMethods();
-        if (logMethods != null && logMethods.contains(method)) {
-            context.getLogger().i("Check method : " + method);
-        }
         Node ownerNode = graph.get(owner);
         if (ownerNode != null) {
             MethodEntity originMethod = ownerNode.confirmOriginMethod(name, desc);
@@ -76,7 +89,7 @@ public class ReferCheckMethodVisitor extends MethodVisitor {
                             if (child instanceof ClassNode && !TypeUtil.isAbstract(child.entity.access)
                                     && TypeUtil.isAbstract(child.confirmOriginMethod(name, desc).access())) {
                                 //非抽象子类没有实现这个抽象类
-                                context.addNotAccessMember(
+                                checkIssueReceiver.addNotAccessMember(
                                         className, methodName, methodDesc, methodAccess, sourceFile, processingLineNumber,
                                         child.entity.name, name, desc, originMethod.access(), InaccessibleNode.TYPE_NOT_IMPLEMENT);
                             }
@@ -89,7 +102,7 @@ public class ReferCheckMethodVisitor extends MethodVisitor {
                         graph.traverseChildren((ClassNode) ownerNode, child -> {
                             if (!TypeUtil.isAbstract(child.entity.access) && TypeUtil.isAbstract(child.confirmOriginMethod(name, desc).access())) {
                                 //非抽象子类没有实现这个抽象方法
-                                context.addNotAccessMember(
+                                checkIssueReceiver.addNotAccessMember(
                                         className, methodName, methodDesc, methodAccess, sourceFile, processingLineNumber,
                                         child.entity.name, name, desc, originMethod.access(), InaccessibleNode.TYPE_NOT_IMPLEMENT);
                             }
@@ -98,17 +111,17 @@ public class ReferCheckMethodVisitor extends MethodVisitor {
                     }
                 }
                 if (!accessible(opcode, originMethod)) {
-                    context.addNotAccessMember(
+                    checkIssueReceiver.addNotAccessMember(
                             className, methodName, methodDesc, methodAccess, sourceFile, processingLineNumber,
                             owner, name, desc, originMethod.access(), InaccessibleNode.TYPE_INACCESS);
                 }
             } else {
-                context.addNotAccessMember(
+                checkIssueReceiver.addNotAccessMember(
                         className, methodName, methodDesc, methodAccess, sourceFile, processingLineNumber,
                         owner, name, desc, 0, InaccessibleNode.TYPE_METHOD_NOT_FOUND);
             }
         } else {
-            context.addNotAccessMember(
+            checkIssueReceiver.addNotAccessMember(
                     className, methodName, methodDesc, methodAccess, sourceFile, processingLineNumber,
                     owner, name, desc, 0, InaccessibleNode.TYPE_CLASS_NOT_FOUND);
         }
@@ -156,19 +169,27 @@ public class ReferCheckMethodVisitor extends MethodVisitor {
             FieldEntity originField = ownerNode.confirmOriginField(name, desc);
             if (originField != null) {
                 if (!accessible(opcode, originField)) {
-                    context.addNotAccessMember(
+                    checkIssueReceiver.addNotAccessMember(
                             className, methodName, methodDesc, methodAccess, sourceFile, processingLineNumber,
                             owner, name, desc, originField.access(), InaccessibleNode.TYPE_INACCESS);
                 }
             } else {
-                context.addNotAccessMember(
+                checkIssueReceiver.addNotAccessMember(
                         className, methodName, methodDesc, methodAccess, sourceFile, processingLineNumber,
                         owner, name, desc, 0, InaccessibleNode.TYPE_FIELD_NOT_FOUND);
             }
         } else {
-            context.addNotAccessMember(
+            checkIssueReceiver.addNotAccessMember(
                     className, methodName, methodDesc, methodAccess, sourceFile, processingLineNumber,
                     owner, name, desc, 01, InaccessibleNode.TYPE_CLASS_NOT_FOUND);
         }
+    }
+
+    public interface CheckIssueReceiver {
+        void addNotAccessMember(String callClassName, String callMethodName, String callMethodDesc, int callMethodAccess, @Nullable String sourceFile, int line,
+                                String memberOwner, String memberName, String memberDesc, int memberAccess,
+                                int type);
+
+        List<InaccessibleNode> getInaccessibleNodes();
     }
 }